% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_get_property.R
\name{o_get_property}
\alias{o_get_property}
\title{Retrieve Omeka S property based on id}
\usage{
o_get_property(property_id, base_url = NULL)
}
\arguments{
\item{property_id}{The numeric identifier of an Omeka S property.}

\item{base_url}{Defaults to \code{NULL}, typically set for the active session with
\code{\link[=o_set]{o_set()}}.}
}
\value{
A list object, with everything returned by the API.
}
\description{
Retrieve Omeka S property based on id
}
\examples{
\dontrun{
  o_get_property(property_id = 12)
}
}
