% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_sitrep.R
\name{o_sitrep}
\alias{o_sitrep}
\title{Situation report}
\usage{
o_sitrep(base_url = NULL)
}
\arguments{
\item{base_url}{Endpoint of the API of the Omeka installation, or base url of
the website. Internally, it is ensured that the "api" suffix is always
appended to the given url.}
}
\value{
Nothing, used for its side effects.
}
\description{
Performs a set of tests for troubleshooting common issues, and provides
informative message.
}
\examples{

o_sitrep()
}
