% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{omekas_response}
\alias{omekas_response}
\title{S7 class representing an Omeka S API response}
\usage{
omekas_response(
  data = list(),
  url = character(0),
  type = character(0),
  retrieved_at = Sys.time()
)
}
\arguments{
\item{data}{A list; the parsed json from the API response.}

\item{url}{Full URL from where the response was retrieved.}

\item{type}{The type of contents or resource retrieved from the API, e.g.
"items", or "properties".}

\item{retrieved_at}{Time when the given resource was retrieved.}
}
\description{
S7 class representing an Omeka S API response
}
