% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_get.R
\name{or_get}
\alias{or_get}
\title{Retrieve results from Omeka S API}
\usage{
or_get(id, type = c("items", "properties"), base_url = NULL)
}
\arguments{
\item{id}{An Omeka S numeric identifier.}

\item{type}{Defaults to "items". Accepted values include "item" and
"properties".}

\item{base_url}{Endpoint of the API of the Omeka installation, or base url of
the website. Internally, it is ensured that the "api" suffix is always
appended to the given url.}
}
\value{
A list object, with attributes detailing the source url, the type of
API endpoint called, and the timestamp of data retrieval.
}
\description{
Retrieve results from Omeka S API
}
\examples{
\dontrun{
  or_get(id = 123, type = "items")
  or_get(id = "123", type = "items") # id can be given as character
  or_get(id = 12, type = "properties")
}
}
