% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_get_item.R
\name{or_get_item}
\alias{or_get_item}
\title{Retrieve Omeka S item based on id}
\usage{
or_get_item(item_id, base_url = NULL)
}
\arguments{
\item{item_id}{The numeric identifier of an Omeka S item.}

\item{base_url}{Defaults to NULL, typically set for the active session with \code{\link[=or_set]{or_set()}}.}
}
\value{
A list object, with everything returned by the API.
}
\description{
Retrieve Omeka S item based on id
}
\examples{
\dontrun{
  or_get_item(id = 123)
}
}
