% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_sitrep.R
\name{or_sitrep}
\alias{or_sitrep}
\title{Situation report}
\usage{
or_sitrep(base_url = NULL)
}
\arguments{
\item{base_url}{Endpoint of the API of the Omeka installation, or base url of
the website. Internally, it is ensured that the "api" suffix is always
appended to the given url.}
}
\value{
Nothing, used for its side effects.
}
\description{
Performs a set of tests for troubleshooting common issues, and provides
informative message.
}
\examples{

or_sitrep()
}
