% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_set.R
\name{o_set}
\alias{o_set}
\title{Sets settings for the current session.}
\usage{
o_set(base_url = NULL)
}
\arguments{
\item{base_url}{Endpoint of the API of the Omeka installation, or base url of
the website. Internally, it is ensured that the "api" suffix is always
appended to the given url.}
}
\value{
Inivisbly, the settings as a list.
}
\description{
Sets settings for the current session.
}
\examples{

o_set(base_url = "https://example.com/api")

base_url <- o_set()[["base_url"]]

base_url
}
