% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{omekas_item}
\alias{omekas_item}
\title{S7 class representing an Omeka S item}
\usage{
omekas_item(
  data = list(),
  url = character(0),
  type = character(0),
  retrieved_at = (function (.data = double(), tz = "") 
 {
     .POSIXct(.data, tz = tz)

    })(),
  id = integer(0)
)
}
\description{
S7 class representing an Omeka S item
}
\section{Slots}{

\describe{
\item{\code{id}}{Numeric, the identifier of the resource requested.}
}}

