% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{omekas_response}
\alias{omekas_response}
\title{S7 class representing an Omeka S API response}
\usage{
omekas_response(
  data = list(),
  url = character(0),
  type = character(0),
  retrieved_at = (function (.data = double(), tz = "") 
 {
     .POSIXct(.data, tz = tz)

    })()
)
}
\description{
S7 class representing an Omeka S API response
}
\section{Slots}{

\describe{
\item{\code{data}}{A list; the parsed json from the API response.}

\item{\code{url}}{Full URL from where the response was retrieved.}

\item{\code{type}}{The type of contents or resource retrieved from the API, e.g.
"items", or "properties".}

\item{\code{retrieved_at}}{Time when the given resrouce was retrieved.}
}}

