% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omekasr-package.R
\docType{package}
\name{omekasr-package}
\alias{omekasr}
\alias{omekasr-package}
\title{omekasr: Interact with the Omeka S API from R}
\description{
Interact with the Omeka S API from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/giocomai/omekasr}
  \item \url{https://giocomai.r-universe.dev/omekasr}
}

}
\author{
\strong{Maintainer}: Giorgio Comai \email{g@giorgiocomai.eu} (\href{https://orcid.org/0000-0002-0515-9542}{ORCID}) [copyright holder]

}
\keyword{internal}
